@@ -1,6 +1,6 @@
 {
   "name": "harbor",
-  "version": "1.9.0",
+  "version": "1.10.0",
   "description": "Harbor UI with Clarity",
   "angular-cli": {},
   "scripts": {
@@ -20,24 +20,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.3",
-    "@angular/common": "^7.1.3",
-    "@angular/compiler": "^7.1.3",
-    "@angular/core": "^7.1.3",
-    "@angular/forms": "^7.1.3",
-    "@angular/http": "^7.1.3",
-    "@angular/platform-browser": "^7.1.3",
-    "@angular/platform-browser-dynamic": "^7.1.3",
-    "@angular/router": "^7.1.3",
-    "@clr/angular": "^1.0.0",
-    "@clr/icons": "^1.0.0",
-    "@clr/ui": "^1.0.0",
+    "@angular/animations": "^8.2.0",
+    "@angular/common": "^8.2.0",
+    "@angular/compiler": "^8.2.0",
+    "@angular/core": "^8.2.0",
+    "@angular/forms": "^8.2.0",
+    "@angular/platform-browser": "^8.2.0",
+    "@angular/platform-browser-dynamic": "^8.2.0",
+    "@angular/router": "^8.2.0",
+    "@clr/angular": "^2.2.0",
+    "@clr/icons": "^2.2.0",
+    "@clr/ui": "^2.2.0",
     "@fortawesome/fontawesome-free": "^5.1.0-4",
-    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^5.1.0",
     "@ngx-translate/core": "^10.0.2",
     "@ngx-translate/http-loader": "^3.0.1",
     "@types/jquery": "^2.0.41",
-    "@webcomponents/custom-elements": "^1.1.3",
+    "@webcomponents/custom-elements": "^1.2.4",
     "bootstrap": "^4.1.1",
     "buffer": "^5.2.1",
     "core-js": "^2.5.4",
@@ -45,26 +44,26 @@
     "jasmine-core": "^3.3.0",
     "jquery": "^3.4.1",
     "mutationobserver-shim": "^0.3.2",
-    "ng-packagr": "^4.5.0",
+    "ng-packagr": "^5.3.0",
     "ngx-clipboard": "^12.0.0",
     "ngx-cookie": "^1.0.0",
-    "ngx-markdown": "^7.1.5",
+    "ngx-markdown": "^8.1.0",
     "popper.js": "^1.14.3",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.2",
     "stream": "^0.0.2",
-    "swagger-ui": "^3.22.1",
+    "swagger-ui": "^3.23.4",
     "ts-helpers": "^1.1.1",
     "tslib": "^1.9.0",
     "types": "^0.1.1",
-    "web-animations-js": "^2.2.1",
-    "zone.js": "^0.8.26"
+    "web-animations-js": "^2.3.2",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular-devkit/build-ng-packagr": "~0.13.8",
-    "@angular/cli": "^7.1.3",
-    "@angular/compiler-cli": "^7.1.3",
-    "@angular/language-service": "^7.1.3",
+    "@angular-devkit/build-angular": "~0.802.0",
+    "@angular-devkit/build-ng-packagr": "~0.802.0",
+    "@angular/cli": "^8.2.0",
+    "@angular/compiler-cli": "^8.2.0",
+    "@angular/language-service": "^8.2.0",
     "@types/core-js": "^0.9.41",
     "@types/jasmine": "^3.3.1",
     "@types/jasminewd2": "~2.0.3",
@@ -86,9 +85,9 @@
     "rollup-plugin-node-resolve": "^3.0.0",
     "rollup-plugin-uglify": "^1.0.1",
     "ts-node": "~5.0.1",
-    "tsickle": "^0.33.1",
+    "tsickle": "^0.36.0",
     "tslint": "~5.9.1",
-    "typescript": "3.1.3",
+    "typescript": "^3.5.3",
     "typings": "^1.5.0",
     "uglify-js": "3.3.18",
     "webdriver-manager": "10.2.5"
